возникали ошибки в стиле кода
